Delay downcasting to int and check for it where it matters

+++ src/com/rtg/ngs/DeduplicatingNStore.java
@@ -31,7 +31,7 @@ public class DeduplicatingNStore implements UptoNStore {
   private final LongIndex mTopNRes;
   private final IntIndex mResultCounts;
   private final int mN;
-  private final int mNumTemplateSeqs;
+  private final long mNumTemplateSeqs;
   private final long mPositionOffset;
 
   /**
@@ -41,7 +41,7 @@ public class DeduplicatingNStore implements UptoNStore {
    * @param templateMaxLength size of largest template sequence
    * @param readMaxLength maximum read length, used to determine how far off template we can go
    */
-  public DeduplicatingNStore(final int numReads, final int numTemplateSeqs, final int n, final long templateMaxLength, long readMaxLength) {
+  public DeduplicatingNStore(final long numReads, final long numTemplateSeqs, final int n, final long templateMaxLength, long readMaxLength) {
     final long length = numReads * (long) n;
 
     mPositionOffset = readMaxLength;
